shared/experiments/geom_embed/requirements.txt
# Core dependencies
torch>=2.0.0
geoopt>=0.5.0
numpy>=1.21.0

# Web API
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0

# Data and configuration
pyyaml>=6.0
scikit-learn>=1.3.0  # For evaluation metrics

# Development
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# Optional: for hyperbolic operations
# hyptorch>=0.1.0  # Alternative to geoopt

# Optional: for visualization
plotly>=5.17.0
matplotlib>=3.7.0